Move fetching functions to new `fetchers` folder
