+++ src/app/roadmap/[roadmapId]/editRoadmap/page.tsx
@@ -1,5 +1,5 @@
 import RoadmapForm from "../../createRoadmap/roadmapForm";
-import getOneRoadmap from "@/functions/getOneRoadmap";
+import getOneRoadmap from "@/fetchers/getOneRoadmap";
 import { getSessionData } from '@/lib/session';
 import { cookies } from 'next/headers';
 import { notFound } from 'next/navigation';
