+++ src/app/roadmap/[roadmapId]/goal/[goalId]/action/createAction/page.tsx
@@ -1,7 +1,7 @@
 import { getSessionData } from "@/lib/session";
 import { cookies } from "next/headers";
 import ActionForm from "./actionForm";
-import getOneRoadmap from "@/functions/getOneRoadmap";
+import getOneRoadmap from "@/fetchers/getOneRoadmap";
 import { notFound } from "next/navigation";
 import accessChecker from "@/lib/accessChecker";
 import { BackButton } from '@/components/redirectButtons';
