+++ src/app/roadmap/[roadmapId]/goal/[goalId]/page.tsx
@@ -1,5 +1,5 @@
 import { notFound } from "next/navigation";
-import getOneRoadmap from "@/functions/getOneRoadmap";
+import getOneRoadmap from "@/fetchers/getOneRoadmap";
 import { cookies } from "next/headers";
 import { getSessionData } from "@/lib/session";
 import accessChecker from "@/lib/accessChecker";
@@ -10,7 +10,7 @@ import Actions from "@/components/tables/actions";
 import Link from "next/link";
 import Image from "next/image";
 import GraphGraph from "@/components/graphs/graphGraph";
-import getOneGoal from "@/functions/getOneGoal";
+import getOneGoal from "@/fetchers/getOneGoal";
 import { Goal, DataSeries } from "@prisma/client";
 import GraphSelector from "@/components/graphs/graphselector/graphSelector";
